+++ R/geom-pointline.r
@@ -22,17 +22,6 @@ NULL
 #' \emph{also} use this, thus lines between x-values are not grouped.
 #' Use \code{aes(group=...)} to enforce a group across x-values.
 #' 
-#' The \strong{tweak} parameter can be used to fine adjust the slope of a line.
-#' If two points are very close horizontally or vertically, something happens
-#' and the line may appear to be pointing in the wrong direction 
-#' (e.g. upwards when the point indicates it should be downwards). I have no
-#' explanation for this, other than it could be a rounding error.
-#' The size of this weaking parameter depends on the physical size of the output.
-#' Good values could be \code{c(0.02, 0.02)} for very small graphs,
-#' \code{c(0.02, 0.06)} for a low, but wide graph, or \code{c(0.06, 0.06)} for
-#' a larger graph.
-#' 
-#' 
 #' @section Aesthetics:
 #' \code{geom_pointline} and \code{geom_pointpath} understands the following 
 #' aesthetics (required aesthetics are in bold):
@@ -74,17 +63,14 @@ NULL
 #'   define both data and aesthetics and shouldn't inherit behaviour from the 
 #'   default plot specification, e.g. \code{\link[ggplot2]{borders}}.
 #' @param linesize Width of of line.
-#' @param distance Gap size between point and end of lines;
+#' @param distance Gap size between point and end of lines when the distance
+#'   between points exceeds \code{threshold};
 #'   use \code{\link[grid]{unit}}. Is converted to 'pt' if given as simple numeric.
+#' @param threshold,shorten Use \code{threshold} to adjust when the line between
+#'   two points should be drawn with a gap (per argument above) or shortened by
+#'   the proportion given by \code{shorten}.  
 #' @param linecolour,linecolor When not \code{waiver()}, the line is drawn with 
 #'   this colour instead of that set by aesthetic \code{colour}.
-#' @param tweak Numeric vector for x or y tweaking parameter.
-#'   When \code{NULL} and \code{NA}, \code{tweak = c(0,0).}
-#' 
-#' @section Known bugs:
-#' Issue #7: When points are too close, the line between them might be 
-#' rendered wrongly.
-#' We have no solution to it currently.
 #' 
 #' @example inst/examples/geom-pointline-ex.r
 #' @export
@@ -92,6 +78,8 @@ geom_pointpath <- function(mapping = NULL, data = NULL, stat = "identity",
                       position = "identity", na.rm = FALSE,
                       show.legend = NA, inherit.aes = TRUE, 
                       distance = unit(3, 'pt'), 
+                      shorten = 0.5,
+                      threshold = 0.1,
                       lineend = "butt",
                       linejoin = "round",
                       linemitre = 1,
@@ -99,7 +87,6 @@ geom_pointpath <- function(mapping = NULL, data = NULL, stat = "identity",
                       linecolour = waiver(),
                       linecolor = waiver(),
                       arrow = NULL,
-                      tweak = c(0.02, 0.03), 
                       ...) {
   if (is.waive(linecolour) && !is.waive(linecolor)) linecolour <- linecolor
   
@@ -114,13 +101,14 @@ geom_pointpath <- function(mapping = NULL, data = NULL, stat = "identity",
     params = list(
       na.rm = na.rm,
       distance = distance,
+      shorten = shorten,
+      threshold = threshold,
       lineend = lineend,
       linejoin = linejoin,
       linemitre = linemitre,
       linesize = linesize,
       linecolour = linecolour,
       arrow = arrow,
-      tweak = tweak,
       ...
     )
   )
@@ -145,25 +133,20 @@ GeomPointPath <- ggplot2::ggproto('GeomPointPath',
   ),
   
   draw_panel = function(data, panel_params, coord, na.rm = FALSE,
-                        distance = grid::unit(3, 'pt'), linesize = 0.5,
+                        distance = grid::unit(3, 'pt'), 
+                        shorten = 0.5,
+                        threshold = 0.1,
+                        linesize = 0.5,
                         linecolour = waiver(),
                         arrow = NULL,
-                        lineend = "butt", linejoin = "round", linemitre = 1,
-                        tweak = c(0.02, 0.03) 
+                        lineend = "butt", linejoin = "round", linemitre = 1
                         ) {
     # Test input parameters
     if (is.null(distance) || is.na(distance)) 
       distance=grid::unit(0, 'pt')
-    if (!is.unit(distance) && is.numeric(distance)) 
+    if (!grid::is.unit(distance) && is.numeric(distance)) 
       distance <- grid::unit(distance, 'pt')
     
-    if (is.null(tweak))
-      tweak = c(0.0,0.0)
-    if (any(is.na(tweak)))
-      tweak[is.na(tweak)] = 0
-    if (length(tweak) == 1)
-      tweak = c(tweak,tweak)
-
     # Contents of GeomPoint$draw_panel in geom-point.r
     coords <- coord$transform(data, panel_params)
     saveRDS(coords, 'tmp.rds')
@@ -214,60 +197,73 @@ GeomPointPath <- ggplot2::ggproto('GeomPointPath',
 
     # Work out grouping variables for grobs
     n <- nrow(munched)
-    #start <- c(TRUE, group_diff)
     group_diff <- munched$group[-1] != munched$group[-n]
-    end <-   c(group_diff, TRUE)
+    munched$end <-   c(group_diff, TRUE)
 
+    munched <- within(munched, {
+      x1 = c(x[-1], NA);
+      y1 = c(y[-1], NA);
+      deltax = x1 - x;
+      deltay = y1 - y;
+      length <- sqrt(deltax**2 + deltay**2);
+    })
     
-    # Make df with x0,y0,x1,y1
-    munched$x1 <- c(munched$x[-1], NA)
-    munched$y1 <- c(munched$y[-1], NA)
-    munched$end <- end
-    
-    # Calculate angle between each pair of points:
-    if (TRUE && as.numeric(distance) != 0) {
-      munched <- within(munched, {
-        end = ifelse(sqrt((x-x1)**2 + (y-y1)**2) < tooshort, TRUE, end);
-        length = sqrt((x-x1)**2 + (y-y1)**2);
-        theta = atan2(y1-y, x1-x);
-        
-        #theta = ifelse(abs(y1 - y) < tweak[2], round(theta / pi * 4) * pi / 4, theta);
-        # -0.2 corresponds to 12 degrees, about 1/16th of a radian
-        # 1.371 = pi/2 - 0.2; 1.771 = pi/2 + 0.2
-        deltay = y1 - y;
-        adjust = abs(y1 - y) < tweak[2] | abs(x1 - x) < tweak[1];
-        if (any(adjust, na.rm=TRUE)) {
-          theta = ifelse(adjust & -pi/4 < theta & theta <= pi/4, pmax(pmin(theta, 0.2), -0.2), theta);
-          theta = ifelse(adjust & pi/4 < theta & theta <= pi*3/4, pmax(pmin(theta, 1.771), 1.371), theta);
-          theta = ifelse(adjust & -pi/3*4 < theta & theta <= -pi/4, pmax(pmin(theta, -1.771), -1.371), theta);
-          theta = ifelse(adjust & theta < -pi*3/4, pmin(-3.351, theta), theta);
-          theta = ifelse(adjust & theta > pi*3/4, pmax(3.351, theta), theta);
-        }
-
-        size = grid::unit(size, 'pt');
-        x = grid::unit(x, 'native') + size*cos(theta) + distance*cos(theta);
-        x1 = grid::unit(x1, 'native') - size*cos(theta) - distance*cos(theta);
-        y = grid::unit(y, 'native') + size*sin(theta) + distance*sin(theta);
-        y1 = grid::unit(y1, 'native') - size*sin(theta) - distance*sin(theta);
-      })
+    if (any(munched$length > threshold)) {
+      # Calculate angle between each pair of points and move endpoints:
+      if (as.numeric(distance) != 0) {
+        df <- within(munched[munched$length > threshold,], {
+          theta = atan2(deltay, deltax);
+          size = grid::unit(size, 'pt');
+          x = grid::unit(x, 'native') + size*cos(theta) + distance*cos(theta);
+          x1 = grid::unit(x1, 'native') - size*cos(theta) - distance*cos(theta);
+          y = grid::unit(y, 'native') + size*sin(theta) + distance*sin(theta);
+          y1 = grid::unit(y1, 'native') - size*sin(theta) - distance*sin(theta);
+        })
+      }
+      
+      gr_distant <- with(df, grid::segmentsGrob(
+        x0=x[!end], y0=y[!end], x1=x1[!end], y1=y1[!end],
+        arrow = arrow,
+        gp = grid::gpar(
+          col = 'green', #ggplot2::alpha(colour, alpha)[!end],
+          fill = ggplot2::alpha(colour, alpha)[!end],
+          lwd = linesize * .pt,
+          lty = linetype[!end],
+          lineend = lineend,
+          linejoin = linejoin,
+          linemitre = linemitre
+        )
+      ))
+      if (!is.waive(linecolour)) 
+        gr_distant$gp$col <- linecolour
     }
     
-    gr_lines <- grid::segmentsGrob(
-      x0=munched$x[!end], y0=munched$y[!end], x1=munched$x1[!end], y1=munched$y1[!end],
-      arrow = arrow,
-      gp = grid::gpar(
-        col = ggplot2::alpha(munched$colour, munched$alpha)[!end],
-        fill = ggplot2::alpha(munched$colour, munched$alpha)[!end],
-        lwd = linesize * .pt,
-        lty = munched$linetype[!end],
-        lineend = lineend,
-        linejoin = linejoin,
-        linemitre = linemitre
-      )
-    )
-    if (!is.waive(linecolour)) gr_lines$gp$col <- linecolour
-
-    grid::gList(gr_points, gr_lines)
+    if (any(munched$length <= threshold)) {
+      df <- within(munched[munched$length <= threshold,], {
+        x  = x + shorten/2 * deltax; x = grid::unit(x, 'native');
+        y  = y + shorten/2 * deltay; y = grid::unit(y, 'native');
+        x1 = x1 - shorten/2 * deltax; x1 = grid::unit(x1, 'native');
+        y1 = y1 - shorten/2 * deltay; y1 = grid::unit(y1, 'native');
+      })
+      
+      gr_short <- with(df, grid::segmentsGrob(
+        x0=x[!end], y0=y[!end], x1=x1[!end], y1=y1[!end],
+        arrow = arrow,
+        gp = grid::gpar(
+          col = 'red', #ggplot2::alpha(colour, alpha)[!end],
+          fill = ggplot2::alpha(colour, alpha)[!end],
+          lwd = linesize * .pt,
+          lty = linetype[!end],
+          lineend = lineend,
+          linejoin = linejoin,
+          linemitre = linemitre
+        )
+      ))
+      if (!is.waive(linecolour)) 
+        gr_short$gp$col <- linecolour
+    }
+    
+    grid::gList(gr_points, gr_distant, gr_short)
   },
   
   draw_key = ggplot2::draw_key_point
@@ -288,7 +284,6 @@ geom_pointline <- function(mapping = NULL, data = NULL, stat = "identity",
                            linecolour = waiver(),
                            linecolor = waiver(),
                            arrow = NULL,
-                           tweak = c(0.02, 0.03), 
                            ...) {
   
   if (is.waive(linecolour) && !is.waive(linecolor)) linecolour <- linecolor
@@ -310,7 +305,6 @@ geom_pointline <- function(mapping = NULL, data = NULL, stat = "identity",
       linesize = linesize,
       linecolour = linecolour,
       arrow = arrow,    
-      tweak = tweak,
       ...
     )
   )